separate unit tests according to new benchmark/assembly structure
