+++ tests/test_metrics/test_ceiling.py
@@ -1,11 +1,11 @@
-import pytest
-from pytest import approx
 from string import ascii_lowercase as alphabet
 
 import numpy as np
-from brainio_base.assemblies import NeuroidAssembly, DataAssembly
+import pytest
+from pytest import approx
 
-from brainscore.assemblies.private import DicarloMajaj2015Loader
+from brainio_base.assemblies import NeuroidAssembly, DataAssembly
+from brainscore.benchmarks.majaj2015 import load_assembly
 from brainscore.metrics.ceiling import NoCeiling, InternalConsistency, SplitHalfConsistency
 
 
@@ -14,6 +14,7 @@ class TestNoCeiling:
         ceiling = NoCeiling()
         ceiling_score = ceiling()
         assert ceiling_score == 1
+
 
 class TestInternalConsistency:
     def test_dummy_data(self):
@@ -30,8 +31,7 @@ class TestInternalConsistency:
 
     @pytest.mark.private_access
     def test_majaj2015_it(self):
-        loader = DicarloMajaj2015Loader('private')
-        assembly_repetitions = loader(average_repetition=False).sel(region='IT')
+        assembly_repetitions = load_assembly(region='IT')
         ceiler = InternalConsistency()
         ceiling = ceiler(assembly_repetitions)
         assert ceiling.sel(aggregation='center') == approx(.82, abs=.01)
